feat: Add `is_fungible` flag to standard definitions
This introduces a new boolean flag, `is_fungible`, to the `[template.fixed]`
section of the voucher standard definition.

This property is crucial for defining whether the nominal values of vouchers
of a certain standard are interchangeable and can be summed up. It allows
client applications (like a wallet) to correctly handle balances.

- Added `is_fungible` to the `TemplateFixed` struct.
- Updated the `standard_template.toml` with documentation.
- Set the flag to `true` for the existing `minuto` and `silver` standards.